Avoid MOVBE instruction, avoid auto-generated AVX2 code.
The NonAvxBitmapFiller must not be in the same .cpp file as the AvxBitmapFiller, because then AVX2 code would be auto-generated.
Use TEMPLATE_TEST_CASE to avoid test code duplication.
