+++ DeepSkyStackerTest/DeepSkyStackerTest.vcxproj
@@ -153,7 +153,6 @@
     <ClCompile Include="..\DeepSkyStacker\avx_cfa.cpp" />
     <ClCompile Include="..\DeepSkyStacker\avx_support.cpp" />
     <ClCompile Include="..\DeepSkyStacker\BitMapFiller.cpp" />
-    <ClCompile Include="AvxBitmapFillerTest.cpp" />
     <ClCompile Include="AvxCfaTest.cpp" />
     <ClCompile Include="BitMapFillerTest.cpp" />
     <ClCompile Include="DeepSkyStackerTest.cpp" />
