+++ DeepSkyStackerTest/DeepSkyStackerTest.vcxproj.filters
@@ -39,9 +39,6 @@
     <ClCompile Include="..\DeepSkyStacker\avx_bitmap_filler.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="AvxBitmapFillerTest.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="catch.h">
